messenger: allow other users to set alarms, removed session edition - autosave mode left user_settings: fixed bug with getting other users setting event: messenger alerts for other users

+++ modules/Base/User/Settings/SettingsCommon_0.php
@@ -114,17 +114,17 @@ class Base_User_SettingsCommon extends ModuleCommon {
 		if (!Acl::is_user()) return null;
 		if ($user===null) $user = Acl::get_user();
 		$module = str_replace('/','_',$module);
-		if (!isset(self::$user_variables))
-			self::$user_variables = array();
-		if (!isset(self::$user_variables[$module])) {
-			self::$user_variables[$module] = array();
+		if (!isset(self::$user_variables[$user]))
+			self::$user_variables[$user] = array();
+		if (!isset(self::$user_variables[$user][$module])) {
+			self::$user_variables[$user][$module] = array();
 			$ret = DB::Execute('SELECT variable, value FROM base_user_settings WHERE user_login_id=%d AND module=%s',array($user, $module));
 			while($row = $ret->FetchRow())
-				self::$user_variables[$module][$row['variable']] = unserialize($row['value']);
+				self::$user_variables[$user][$module][$row['variable']] = unserialize($row['value']);
 		}
-		if (isset(self::$user_variables[$module][$name]))
-			return self::$user_variables[$module][$name];	
-		return self::$user_variables[$module][$name] = self::get_admin($module,$name);
+		if (isset(self::$user_variables[$user][$module][$name]))
+			return self::$user_variables[$user][$module][$name];	
+		return self::$user_variables[$user][$module][$name] = self::get_admin($module,$name);
 	}
 
 	/**
@@ -136,17 +136,19 @@ class Base_User_SettingsCommon extends ModuleCommon {
 	 * @param mixed value
 	 * @return bool true on success, false otherwise
 	 */
-	public static function save($module,$name,$value){
+	public static function save($module,$name,$value,$user=null){
 		if (!Acl::is_user()) return false;
 		//if ($value === null) $value = 0;
 		$module = str_replace('/','_',$module);
 		$def = self::get_admin($module,$name);
 //		if (!isset($def)) return false;
+		if (!Acl::is_user()) return null;
+		if ($user===null) $user = Acl::get_user();
 		if ($value==$def) {
 			DB::Execute('DELETE FROM base_user_settings WHERE user_login_id=%d AND module=%s AND variable=%s',array(Acl::get_user(),$module,$name));
-			if(isset(self::$user_variables)) unset(self::$user_variables[$module][$name]);
+			if(isset(self::$user_variables[$user])) unset(self::$user_variables[$user][$module][$name]);
 		} else {
-			if(isset(self::$user_variables)) self::$user_variables[$module][$name]=$value;
+			if(isset(self::$user_variables[$user])) self::$user_variables[$user][$module][$name]=$value;
 			$value = serialize($value);
 			$val = DB::GetOne('SELECT value FROM base_user_settings WHERE user_login_id=%d AND module=%s AND variable=%s',array(Acl::get_user(),$module,$name));
 			if ($val === false || $val===null)
